# openapi/ao-os.openapi.yaml
openapi: 3.1.0
info:
  title: AO OS API
  version: 0.2.0
  description: |
    Operations platform API for AO Sanctuary.
    All endpoints require Bearer JWT authentication unless noted.
    Roles: `front_desk`, `operations`, `admin`.
servers:
  - url: https://api.aosanctuary.com/v1
    description: Production
  - url: http://localhost:4000/v1
    description: Local development
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Guest:
      type: object
      properties:
        id: { type: string, format: uuid }
        first_name: { type: string }
        last_name: { type: string, nullable: true }
        phone: { type: string, nullable: true }
        email: { type: string, nullable: true }
        membership_status:
          type: string
          enum: [none, basic, premium]
        risk_flag_status:
          type: string
          enum: [clear, flagged, banned]
        version: { type: integer }
        created_at: { type: string, format: date-time }

    # ── Waivers ─────────────────────────────────────────────────────────────
...
paths:
...
  /guests:
    post:
      summary: Create a guest
      tags: [Guests]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [first_name]
              properties:
                first_name: { type: string }
                last_name: { type: string }
                phone: { type: string }
                email: { type: string, format: email }
                date_of_birth: { type: string, format: date }
      responses:
        '201':
          description: Guest created
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Guest' }
  /guests/{guestId}:
    get:
      summary: Get a guest by ID
      tags: [Guests]
      parameters:
        - name: guestId
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Guest found
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Guest' }
        '404':
          description: Not found

    patch:
      summary: Update a guest
      tags: [Guests]
      parameters:
        - name: guestId
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                first_name: { type: string }
                last_name: { type: string }
                phone: { type: string }
                email: { type: string, format: email }
      responses:
        '200':
          description: Guest updated
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Guest' }

  # ══════════════════════════════════════════════════════════════════════════
  # Waivers
  # ══════════════════════════════════════════════════════════════════════════
...